refactor(cli): extract config loading utilities into shared helper
