+++ lib/python/openlinktoken-cli/src/main/openlinktoken_cli/commands/package_command.py
@@ -11,14 +11,9 @@ from openlinktoken.metadata import Metadata
 from openlinktoken.tokentransformer.encrypt_token_transformer import EncryptTokenTransformer
 from openlinktoken.tokentransformer.hash_token_transformer import HashTokenTransformer
 from openlinktoken.tokentransformer.token_transformer import TokenTransformer
-from openlinktoken_cli.tokens.config.dynamic_attribute_factory import DynamicAttributeFactory
-from openlinktoken_cli.tokens.config.dynamic_token_definition import DynamicTokenDefinition
-from openlinktoken_cli.tokens.config.tokenization_config import TokenizationConfig
-from openlinktoken_cli.tokens.config.tokenization_config_loader import TokenizationConfigLoader
-from openlinktoken_cli.io.csv.person_attributes_csv_reader import PersonAttributesCSVReader
+from openlinktoken_cli.commands.tokenization_config_helper import TokenizationConfigHelper
 from openlinktoken_cli.io.csv.person_attributes_csv_writer import PersonAttributesCSVWriter
 from openlinktoken_cli.io.json.metadata_json_writer import MetadataJsonWriter
-from openlinktoken_cli.io.parquet.person_attributes_parquet_reader import PersonAttributesParquetReader
 from openlinktoken_cli.io.parquet.person_attributes_parquet_writer import PersonAttributesParquetWriter
 from openlinktoken_cli.processor.person_attributes_processor import (
     PersonAttributesProcessingSummary,
@@ -181,7 +176,7 @@ class PackageCommand:
                     # Determine total rows via reader to enable %/ETA
                     total_rows: int | None = None
                     try:
-                        reader = PackageCommand._create_reader(
+                        reader = TokenizationConfigHelper.create_reader(
                             args.input_path, input_type)
                         total_rows = reader.row_count()
                         reader.close()
@@ -274,9 +269,9 @@ class PackageCommand:
             raise RuntimeError("Failed to initialize transformers") from e
 
         try:
-            config, factory, token_definition = PackageCommand._load_tokenization_config(tokenization_config_path)
+            config, factory, token_definition = TokenizationConfigHelper.load_tokenization_config(tokenization_config_path)
             with (
-                PackageCommand._create_reader(input_path, input_type, config, factory) as reader,
+                TokenizationConfigHelper.create_reader(input_path, input_type, config, factory) as reader,
                 PackageCommand._create_writer(output_path, output_type) as writer,
             ):
                 # Create metadata
@@ -331,56 +326,6 @@ class PackageCommand:
         if hash_record_ids:
             lines.append("Record ID hashing: enabled")
         return lines
-
-    @staticmethod
-    def _create_reader(
-        path: str,
-        file_type: str,
-        config: TokenizationConfig = None,
-        factory: DynamicAttributeFactory = None,
-    ):
-        """Create a PersonAttributesReader based on file type."""
-        attribute_map = None
-        if config and factory:
-            attribute_map = PackageCommand._build_configured_input_attribute_map(config, factory)
-
-        file_type_lower = file_type.lower()
-        if file_type_lower == FileTypeDetector.TYPE_CSV:
-            if attribute_map is not None:
-                return PersonAttributesCSVReader(path, attribute_map=attribute_map)
-            return PersonAttributesCSVReader(path)
-        elif file_type_lower == FileTypeDetector.TYPE_PARQUET:
-            if attribute_map is not None:
-                return PersonAttributesParquetReader(path, attribute_map=attribute_map)
-            return PersonAttributesParquetReader(path)
-        else:
-            raise ValueError(f"Unsupported input type: {file_type}")
-
-    @staticmethod
-    def _load_tokenization_config(
-        tokenization_config_path: str = None,
-    ) -> tuple[TokenizationConfig | None, DynamicAttributeFactory | None, DynamicTokenDefinition | None]:
-        if not tokenization_config_path:
-            return None, None, None
-
-        config = TokenizationConfigLoader.load(tokenization_config_path)
-        factory = DynamicAttributeFactory(config)
-        token_definition = DynamicTokenDefinition(config, factory)
-        return config, factory, token_definition
-
-    @staticmethod
-    def _build_configured_input_attribute_map(
-        config: TokenizationConfig,
-        factory: DynamicAttributeFactory,
-    ) -> dict:
-        """Build explicit column-to-dynamic-attribute mapping for config-driven reads."""
-        attribute_map = {}
-        for csv_column in config.attributes:
-            try:
-                attribute_map[csv_column] = factory.get_class_for_csv_column(csv_column)
-            except KeyError:
-                logger.warning("CSV column '%s' is in config but has no dynamic class registered.", csv_column)
-        return attribute_map
 
     @staticmethod
     def _create_writer(path: str, file_type: str):
